Silence the MSVC /W4 narrowing warning on the header uppercase bound
sizeof is 64-bit size_t; the parameter is uint32_t. Behaviour-identical
(the bound is 44); all nine sheets re-proof byte-identical.

+++ Engine/Editor/OutlinerPanel.cpp
@@ -1155,7 +1155,7 @@ void OutlinerPanel::RecordFolderRow(EditorInstance* Instances, uint32_t Instance
     }
 
     char Upper[44] = {};
-    UpperCopy(Upper, sizeof(Upper), Row.Label);
+    UpperCopy(Upper, static_cast<uint32_t>(sizeof(Upper)), Row.Label);
     ImGui::PushFont(Small);
     const ImVec2 LabelGlyph = Small->CalcTextSizeA(Small->LegacySize, FLT_MAX, 0.0f, Upper);
     ImGui::PopFont();
